xml_parse_token_map {xmlparsedata}	R Documentation

Map token names of the R parser to token names in xml_parse_data

Description

Some of the R token names are not valid XML tag names, so xml_parse_data needs to replace them to create a valid XML file.

Usage

xml_parse_token_map

Format

An object of class character of length 20.

See Also

xml_parse_data

[Package xmlparsedata version 1.0.5 Index]